refactor: rename abstractDrug to drug
